api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.0
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.7
Pillow>=10.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
